Split Client and ClientBuilder. Remove LoggingBuilder
