+++ Samples/Container/index.ts
@@ -12,7 +12,6 @@ import { MyEventHandler } from './MyEventHandler';
 const client = Client
     .forMicroservice('7a6155dd-9109-4488-8f6f-c57fe4b65bfb')
     .withContainer(Container)
-    .withLogging(_ => _.useWinston(_ => _.level = 'debug'))
     .withEventTypes(eventTypes =>
         eventTypes.register(MyEvent))
     .withEventHandlers(eventHandlers =>
